requirements.txt
altgraph==0.17.4
anyio==4.6.0
astor==0.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
flatbuffers==24.3.25
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imgaug==0.4.0
kiwisolver==1.4.7
lanms_neo==1.0.2
lazy_loader==0.4
lmdb==1.5.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnxruntime-gpu==1.12.1
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==24.1
paddlepaddle==2.6.2
pefile==2024.8.26
pillow==10.4.0
Polygon3==3.0.9.1
protobuf==3.20.2
pyclipper==1.3.0.post5
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
pyreadline3==3.5.4
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
